Updated README and new pip-install-requirements.sh

+++ inject-logging-into-gunicorn.sh
@@ -6,9 +6,10 @@ E=/docker-entrypoint-ext.d
 INI=$E/gunicorn.ini
 L=$E/gunicorn.logging.conf
 
-grep -E '^logconfig' $INI || echo 'logconfig = "$L"' >> $INI
-
-[[ -f $L ]] || cp /gunicorn.logging.conf $L
+if [[ -f $INI ]] ; then
+	grep -E '^logconfig' $INI || echo 'logconfig = "$L"' >> $INI
+	[[ -f $L ]] || cp /gunicorn.logging.conf $L
+fi
